add users' views to views.py from folder views
